# openapi.yaml
openapi: 3.1.0
info:
  title: Cloudflare Live Auction API
  version: 1.0.0
  description: Strongly ordered live auctions backed by one authoritative Durable Object per auction
    and sharded realtime fanout.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT subject becomes actorId; the signed role claim determines seller, bidder, or viewer
        permissions.
  schemas:
...
    DemoBidderSessionSuccess:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - true
        auctionId:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,100}$
          example: vintage-camera-2026
        bidderId:
          type: string
          minLength: 1
          maxLength: 200
        expiresAt:
          type: integer
          exclusiveMinimum: 0
        bidderToken:
          type: string
          minLength: 1
      required:
        - ok
        - auctionId
        - bidderId
        - expiresAt
        - bidderToken
      additionalProperties: false
    Error:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            status:
              type: integer
              minimum: 400
              maximum: 599
          required:
            - code
            - message
            - status
          additionalProperties: false
      required:
        - ok
        - error
      additionalProperties: false
    DemoSessionSuccess:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - true
        auctionId:
          type: string
          pattern: ^[A-Za-z0-9_-]{1,100}$
          example: vintage-camera-2026
        expiresAt:
          type: integer
          exclusiveMinimum: 0
        sellerToken:
          type: string
          minLength: 1
        bidderToken:
          type: string
          minLength: 1
        viewerToken:
          type: string
          minLength: 1
      required:
        - ok
        - auctionId
        - expiresAt
        - sellerToken
        - bidderToken
        - viewerToken
      additionalProperties: false
...
paths:
...
  /v1/demo-session/{auctionId}/bidder:
    post:
      summary: Join a live demo as a new bidder
      description: Available only in demo-enabled environments. Returns a distinct 15-minute bidder token
        scoped to an existing live demo auction.
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
      responses:
        "201":
          description: Demo bidder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoBidderSessionSuccess"
        "404":
          description: Live demo auction not found or demos disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Demo auction is not live
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Demo signing is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/demo-session:
    post:
      summary: Create a short-lived demo room
      description: Available only in demo-enabled environments. Returns 15-minute role tokens scoped to
        one generated auction ID.
      responses:
        "201":
          description: Demo session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoSessionSuccess"
        "404":
          description: Demo sessions are disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Demo signing is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auctions/{auctionId}:
    put:
      summary: Create an auction
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuctionInput"
      responses:
        "200":
          description: Existing identical auction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandSuccess"
        "201":
          description: Auction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role or auction ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Auction state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: Read current auction state
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
      responses:
        "200":
          description: Current auction state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role or auction ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Auction state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auctions/{auctionId}/history:
    get:
      summary: Read ordered auction events
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
        - schema:
            type:
              - number
              - "null"
            default: 0
          required: false
          name: afterSequence
          in: query
        - schema:
            type: number
            default: 50
          required: false
          name: limit
          in: query
      responses:
        "200":
          description: Ordered event page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistorySuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role or auction ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Auction state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auctions/{auctionId}/start:
    post:
      summary: Start an auction
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
        - schema:
            type: string
            pattern: ^Bearer [^\s]+$
          required: true
          name: authorization
          in: header
        - schema:
            type: string
            pattern: ^[\x21-\x7E]{1,200}$
            example: bid-018f4e9f
          required: true
          name: idempotency-key
          in: header
      responses:
        "200":
          description: Command accepted or replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role or auction ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Auction state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auctions/{auctionId}/close:
    post:
      summary: Close an ended auction
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
        - schema:
            type: string
            pattern: ^Bearer [^\s]+$
          required: true
          name: authorization
          in: header
        - schema:
            type: string
            pattern: ^[\x21-\x7E]{1,200}$
            example: bid-018f4e9f
          required: true
          name: idempotency-key
          in: header
      responses:
        "200":
          description: Command accepted or replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role or auction ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Auction state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/auctions/{auctionId}/cancel:
    post:
      summary: Cancel an auction
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^[A-Za-z0-9_-]{1,100}$
            example: vintage-camera-2026
          required: true
          name: auctionId
          in: path
        - schema:
            type: string
            pattern: ^Bearer [^\s]+$
          required: true
          name: authorization
          in: header
        - schema:
            type: string
            pattern: ^[\x21-\x7E]{1,200}$
            example: bid-018f4e9f
          required: true
          name: idempotency-key
          in: header
      responses:
        "200":
          description: Command accepted or replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandSuccess"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient role or auction ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Auction state conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...